Refactor Beard.Date into a lib, add fixed price
